requirements.txt
gradio>=4.37.2
transformers>=4.44.0
torch>=2.3.1
sentence-transformers>=3.0.1
keybert>=0.8.5
scikit-learn>=1.4.2
accelerate>=0.33.0
safetensors
sentencepiece
openai>=1.0.0
spotipy>=2.23.0
python-dotenv>=1.0.1

##가상환경 만들어서 하는걸 추천